Moved activation logic into ElasticContext base class
